feat: wildcards/callable function for container url (#3590)
<!--Add a description of your PR here-->
As the name suggest, the goal of this PR is to be able to change the
container url used for a rule based on the rule's `wildcards`. I have
also allowed for a `callable` to be used as entry for url (change
requested in an [old
issue](https://github.com/snakemake/snakemake/issues/2422#issue-1865707531)
).

This change also allow to sync the container entry standard with the
conda entry standard. The conda keyword already allow for the user to
use `wildcards` and `callable` within a rule but that's not the case for
container and env_modules, which can be a bit confusing imo.

NB: I have also try to add the same change for env_modules but with no
success. If anyone managed to do it, I think it could be a good
addition.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Clarified that container image specifications can be dynamic, allowing
the use of wildcards or callable functions for container image
resolution in workflow rules.
- **New Features**
- Enabled dynamic container image selection based on wildcards or
functions in workflow rules, improving flexibility for containerized job
execution.
- **Tests**
- Added a new test workflow and automated test to validate dynamic
container image specification and wildcard expansion in container
directives.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->