fix: DeprecationWarning when using snakemake.utils.validate (#3420)
This PR closes #2468, by replacing RefResolver functionality with the
referencing library.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Modernized the data validation process by adopting the latest features
of the JSON Schema Draft 2020-12 specification for improved reliability
and consistency.
- Added schema versioning information to existing schema definitions for
enhanced validation and interoperability.
- **Style**
- Enhanced the formatting of warning messages for clearer and more
readable notifications.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->