feat: Prefer papermill to nbconvert (#2857)
As described in #348, the current notebook integration has some
downsides, and they are due to limitations in the `nbconvert` tool.
[`papermill`](https://papermill.readthedocs.io/en/latest/index.html) is
a more feature-rich alternative. In particular, it outputs the executed
notebook incrementally for progress tracking, and outputs the notebook
(with errors) even when it fails.

### QC
* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

I didn't see `nbconvert` anywhere in the docs, so I think this was
always an implementation detail.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Tests**
- Added new test functions for verifying Jupyter notebook execution in
various scenarios.

- **Documentation**
- Updated guidance to clarify headless notebook execution and
expectations during errors.

- **Refactor**
- Enhanced error handling and improved command execution logic for more
reliable Jupyter notebook processing.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Christian Meesters <cmeesters@users.noreply.github.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>