docs: Updated Requirement Section of The Tutorial With Warning of Not Installing The Tools Manually (#3607)
Updated requirement section docs in the tutorial, to guide the user to
wait for preferred ways of installation of required tools to appear in
the following sections and not to install the tools one by one by
yourself through the terminal manually etc.

This was important as some users could do that as the warning may be
hidden as in the original version of docs and secondly had grammatical
mistake.

Fixes #3582

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Improved setup instructions in the tutorial to more clearly advise
users to wait for guided installation steps, enhancing clarity and
reducing confusion.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: rohan-ibn-tariq <your-username@users.noreply.github.com>