fix: Unclear handling of params overriding with inheritance (#3624)
### Description

Fixes #3570 : Overwriting params for rule inheritance did not work
properly when the parent rule did not define a params directive at all.

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Improved handling of rule parameters to prevent errors when parameters
are missing or set to None.

- **Tests**
- Added new tests to verify correct behavior when inheriting and
overriding rule parameters.
  - Included expected output files to validate test results.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->