fix: dag mermaid-js edges (bug:#3556) (#3669)
I noticed that `DAG._dot()` renders the edges correctly, but
`DAG._mermaid_js()` makes a mess out of it.

- I copied the bit from the _dot part to the _mermaid_js part
- modified the f"" stretch to use the correct numbers

This fixed the mermaid output so it matches the dot one.

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Bug Fixes**
* Corrected the direction of edges in the Mermaid.js graph
representation, ensuring dependencies are accurately visualized.
* **New Features**
  * Added a new test workflow to validate complex dependency behaviors.
  * Introduced a visual DAG representation for the new test workflow.
* Added a test case to verify the updated DAG generation with Mermaid.js
output.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->