Add mtime rerun trigger condition in dag.py
Previously, a rule would not be rerun even with newer input files
(according to file timestamps) if the checksums were unchanged. This may
be unexpected and problematic especially for empty "touch" files. Now,
if the mtime rerun trigger is set, checksums are not checked. Because
the mtime rerun trigger is set by default, this also changes default
behaviour. For the old behaviour, only the other rerun-triggers need
to be set ("--rerun-triggers code input params software-env").
Also added a test case for these two conditions (without mtime rerun
trigger and with).