fix: silence messages on module load (#3688)
<!--Add a description of your PR here-->

On some HPC, `module load` prints to `stdout` when loading the
requirements of the environmental modules:
```
$ module load bowtie2/2.5.4 
Loading bowtie2/2.5.4
  Loading requirement: binutils/2.40 gcc/13.2.0
```

Not sure why, but this triggers an error when `snakemake` tries to parse
it:
```
Unable to determine Python version from output 'Loading bowtie2/2.5.4
  Loading requirement: binutils/2.40 gcc/13.2.0
[3, 13]
': Expecting value: line 1 column 1 (char 0)
```

Changes have been tested on both `lmod` and `envmod`.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Refactor**
* Updated environment module loading to suppress output during module
load operations for a cleaner user experience.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->