fix: raise an error when different rules produce identical (temp) output  (#3667)
I noticed an edge case where unused temporary output of different rules
would go to the same file. This can introduce nasty side effects, and
also made it so that my rules needed to be re-run all the times (because
two rules that produce the same output is always detected as a code
change by one of the two rules). This is now nicely raises an exception
giving an early warning to the user that sth is misconfigured.

I'm not sure how this affects performance or other stuff, let me know if
sth needs changes or if this is actually wanted behaviour :smile: