feat: add landing page to the report containing custom metadata defined with a YTE yaml template (#3452)
<!--Add a description of your PR here-->

Adds a landing page for the report, that contains metadata. The landing
page can be reached from every page in the report by pressing the home
button in the navbar on the top left.
The metadata can be user defined by passing a YTE yaml template with
`--report-metadata`. If no metadata template is provided by the user,
only the Snakemake version that was used to run the workflow is
displayed on the landing page.

Related to issue #3395 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced report customization by allowing users to supply a YAML
template for additional metadata.
- Improved the report interface to display enriched, user-defined
contextual information.
- Added a new command-line option to specify custom metadata for
reports.
  
- **Documentation**
- Added a new section with clear examples and instructions for
incorporating custom metadata into reports.
  
- **Tests**
- Introduced automated tests to ensure that the metadata integration
functions as expected.
- Added new test cases covering standard and nested metadata scenarios.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>