docs: output files within output directories is an error (#2848) (#2913)
<!--Add a description of your PR here-->
Closes #2848, by providing the documentation suggested, and fixes a bug
found while investigating:

The `check_directory_outputs()` function relies on the sorting of file
paths to detect nested outputs. This breaks on windows as file paths are
separated with `\\` rather than `/`, which is sorted after alphanumeric
characters, rather than before.

I've fixed this, and provided a test case, but as there's not currently
a mechanism for distinguishing failures before/during execution (before
being the intended behavior, as this provides better feedback to the
user), it would pass either way. A better test would be the code below,
but I don't want to stray from the use of `run()` in the other test
cases. It may also be possible to extend run with an
`expected_exception` parameter for non-shell execution.

```
def test_github_issue2848():
    # Should fail on parsing with ChildIOException, on Windows previously failed during execution
    with api.SnakemakeApi() as snakemake_api:

        workflow_api = snakemake_api.workflow(
            resource_settings=settings.ResourceSettings(cores=1),
            snakefile=Path(dpath("test_github_issue2848/Snakefile")),
            workdir=Path(dpath("test_github_issue2848")),
        )

        dag_api = workflow_api.dag(dag_settings=settings.DAGSettings(targets={"output/c.txt", "output/dir2/d.txt"}))

        with pytest.raises(ChildIOException):
            dag_api.execute_workflow()
```

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- Documentation
• Updated the docs with a sidebar note explaining that directories
marked as outputs are removed before a job runs, helping prevent
accidental file deletion.
  
- New Features
• Added an example workflow that demonstrates file and directory
management in job execution.

- Tests
• Expanded test coverage to ensure improved error handling around
workflow output parsing.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>