fix: fail if single-file report is generated with output directories marked for inclusion (#3748)
### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Added early validation for embedded HTML reports to detect auxiliary
files and raise a clear error directing users to use a ZIP report
instead, preventing late failures during rendering.

- **Tests**
- Added a functional test confirming that providing a report directory
where a file is expected fails as intended.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->