fix: refactor LoggerManager setup and scope (#3851)
Currently there is a single global `LoggerManager` instance, with its
`setup()` method called from `SnakemakeApi`'s constructor. This is a
potential source of issues if the global instance gets used multiple
times in the same Python session, as happens when running the test
suite. It's also not very necessary, as the global is only used in a
couple of places (`api.py` and `workflow.py`), and there's not really
any setup or configuration that actually needs to be shared (a couple of
exceptions are mentioned below).

This PR removes the global instance and instead creates a new one
attached to each `SnakemakeApi` instance. This instance is also now
passed explicitly to `Workflow`'s constructor. Because the setup now
happens when the `LoggerManager` is created, I removed the `setup()`
function and merged its contents into `__init__()`.

I also refactored and simplified some of the code previously in
`setup()`, and also moved the initial creation of plugin log handlers
from `SnakemakeApi.setup_logger()` to `LoggerManager.__init__()`.

Edit: Also removed the `cleanup_logfile()` method and instead remove all
handlers from the logger in `stop()`.

### Potential issues

#### No tests

Currently there are no tests for most of this behavior, that is fixed in
#3848. I have personally tested that those tests still succeed with
these changes, but probably best to merge that PR first.

#### Global logger

There is still a single global `Logger` instance, which gets (mostly)
reset and then set up again for each `LoggerManager` instance. Ideally
each `LoggerManager` would create its own logger, however this is much
more difficult to refactor because the global logger is used extensively
throughout the codebase. I'll leave that to a separate PR/issue.

#### `OutputSettings.keep_logger`

Setup and teardown of the global `LoggerManager` was previously skipped
if `OutputSettings.keep_logger` was True. As far as I can tell this was
never the case - the dataclass field's default value is False, I checked
all instances I could find of the constructor being called and the only
case that sets the field is in
`snakemake/cli.py:create_output_settings()`, where it is always set to
False
([link](https://github.com/snakemake/snakemake/blob/12ce6c6cb854867cf90255c4119b73614a1ec58a/src/snakemake/cli.py#L1923)).
I removed the attribute from `OutputSettings`. If we want to optionally
allow for similar behavior, I could add an argument to `SnakemakeApi`'s
constructor that accepts an existing instance.

#### Default stream handler level

In current `LoggerManager.setup()`:

```python3
class LoggerManager:
	def setup(...):
		...
	
        # Skip plugin handlers if requested and return early
        if settings.skip_plugin_handlers or not handlers:
        
            handler = self._default_streamhandler()
            if settings.log_level_override is not None:
                handler.setLevel(settings.log_level_override)
            self.logger.addHandler(handler)
			
			...
			
			return
```

I removed this if statement because most of the code was duplicated
below (some of it is conditional on no plugin handlers outputting to
stream, which is always the case if there are no handlers or plugins are
skipped). The only thing that is different is that the default stream
handler's level is set based on `log_level_override`, which doesn't
happen otherwise. I'm pretty confident this has no effect because the
`Logger` the handler is attached to is set to the same level.



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Refactor**
* Centralized logging into a manager with plugin-capable initialization,
unified lifecycle and cleanup; logging context is now passed into
workflow execution and logfile handling is managed centrally.
* **Chores**
* Marked the previous keep_logger setting as deprecated and removed its
use during output-settings creation to simplify logging configuration.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Cade Mirchandani <cmirchan@ucsc.edu>